+++ controler/ConnexionControler.class.php
@@ -212,6 +212,7 @@ class ConnexionControler extends PastellControler
         try {
             $config = $this->getConnecteurFactory()->getGlobalConnecteurConfig('message-oublie-identifiant');
         } catch (Exception $e) {
+            /* Nothing to do */
         }
         
     
@@ -228,6 +229,9 @@ class ConnexionControler extends PastellControler
     public function changementMdpAction()
     {
         $recuperateur = new Recuperateur($_GET);
+        $this->{'login_page_configuration'} = file_exists(LOGIN_PAGE_CONFIGURATION_LOCATION)
+            ? file_get_contents(LOGIN_PAGE_CONFIGURATION_LOCATION)
+            : '';
         $this->{'mail_verif_password'} = $recuperateur->get('mail_verif');
         $this->{'page'} = "oublie_identifiant";
         $this->{'page_title'} = "Oubli des identifiants";
@@ -399,15 +403,18 @@ class ConnexionControler extends PastellControler
         $id_u = $utilisateurListe->getByVerifPassword($mail_verif_password);
 
         if (! $id_u) {
-            $this->setLastError("Utilisateur inconnu");
-            $this->redirect("/Connexion/index");
+            /* Note : on ne peut pas mettre de message d'erreur personnalisé pour le moment */
+            echo "Le lien du mail a expiré. Veuillez recommencer la procédure";
+            exit_wrapper();
         }
 
         if (! $password) {
+            /* Note : on ne peut pas mettre de message d'erreur personnalisé pour le moment */
             $this->setLastError("Le mot de passe est obligatoire");
             $this->redirect("/Connexion/changementMdp?mail_verif=$mail_verif_password");
         }
         if ($password != $password2) {
+            /* Note : on ne peut pas mettre de message d'erreur personnalisé pour le moment */
             $this->setLastError("Les mots de passe ne correspondent pas");
             $this->redirect("/Connexion/index");
             $this->redirect("/Connexion/changementMdp?mail_verif=$mail_verif_password");
@@ -424,8 +431,9 @@ class ConnexionControler extends PastellControler
         $utilisateur->reinitPassword($id_u, $mailVerifPassword);
 
         $this->getJournal()->add(Journal::MODIFICATION_UTILISATEUR, $infoUtilisateur['id_e'], 0, "mot de passe modifié", "{$infoUtilisateur['login']} ({$infoUtilisateur['id_u']}) a modifié son mot de passe");
-        $this->setLastMessage("Votre mot de passe a été modifié");
 
+        /* Note : on ne peut pas mettre de message d'erreur personnalisé pour le moment */
+        $this->setLastMessage("Votre mot de passe a été modifié");
         $this->redirect("/Connexion/index");
     }
 
@@ -472,7 +480,7 @@ class ConnexionControler extends PastellControler
         );
 
         $this->setLastMessage("Un email vous a été envoyé avec la suite de la procédure");
-        $this->redirect("/Connexion/index");
+        $this->redirect("/Connexion/oublieIdentifiant");
     }
 
     public function indexAction()