+++ CHANGELOG.md
@@ -25,6 +25,7 @@ l'onglet du parapheur n'apparaissait jamais #977
 - L'identifiant du bordereau de signature passe de `bordereau` à `bordereau_signature` pour tous les types de dossier
 studio, le champ faisait doublon avec le bordereau d'acquittement #987
 - studio : Les actes ne pouvaient pas être annulés #988
+- La page de changement de mot de passe (avant connexion) n'était pas charté #1002
 
 # [3.0.1] - 2019-11-18